src/recommendationservice/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
backoff==1.10.0
    # via opentelemetry-exporter-otlp-proto-grpc
cachetools==4.2.0
    # via google-auth
certifi==2020.12.5
    # via requests
charset-normalizer==2.0.10
    # via requests
deprecated==1.2.13
    # via opentelemetry-api
google-api-core==2.4.0
    # via -r requirements.in
google-auth==2.5.0
    # via google-api-core
googleapis-common-protos==1.52.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio==1.43.0
    # via
    #   -r requirements.in
    #   grpcio-health-checking
    #   opentelemetry-exporter-otlp-proto-grpc
grpcio-health-checking==1.43.0
    # via -r requirements.in
idna==2.10
    # via requests
opentelemetry-api==1.9.1
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc==1.9.1
    # via -r requirements.in
opentelemetry-instrumentation==0.28b1
    # via
    #   -r requirements.in
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-urllib3
opentelemetry-instrumentation-grpc==0.28b1
    # via -r requirements.in
opentelemetry-instrumentation-urllib3==0.28b1
    # via -r requirements.in
opentelemetry-proto==1.9.1
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.9.1
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation-grpc
opentelemetry-semantic-conventions==0.28b1
    # via
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-sdk
opentelemetry-util-http==0.28b1
    # via opentelemetry-instrumentation-urllib3
protobuf==3.13.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   opentelemetry-proto
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
python-json-logger==2.0.2
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   google-api-core
rsa==4.6
    # via google-auth
six==1.15.0
    # via
    #   google-auth
    #   grpcio
    #   protobuf
typing-extensions==4.0.1
    # via opentelemetry-sdk
urllib3==1.26.8
    # via
    #   -r requirements.in
    #   requests
wrapt==1.13.3
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-grpc
    #   opentelemetry-instrumentation-urllib3

# The following packages are considered to be unsafe in a requirements file:
# setuptools
python-consul==1.1.0
